switched back to JpaRepository
* added hibernate dialiect for MySQL
* minor refactoring: moved service interface and implementation into the
  same package
